refactor(core,react): name the origin lock requireOrigin, add usePikkuAnalytics
Nothing in the middleware is analytics-specific — it is an origin lock for any
unauthed route, and analytics was only its first caller. `allowedOrigin` was the
other candidate but sits two characters from the `isAllowedOrigin` predicate
exported beside it.

`usePikkuAnalytics` reaches a `createAnalytics` client through PikkuProvider, so
components consume it the same way as `usePikkuFetch` and `usePikkuRPC` rather
than importing an app module singleton.
