+++ packages/cli/src/functions/wirings/middleware/serialize-middleware-types.ts
@@ -29,7 +29,7 @@ import type { PikkuAgentMiddlewareHooks } from '@pikku/core/agent'
 import type { Services, SingletonServices } from '${functionTypesImportPath}'
 ${requiredServicesTypeImport}
 
-export { cors, analyticsOrigin, authAPIKey, authBearer, authCookie } from '@pikku/core/middleware'
+export { cors, requireOrigin, authAPIKey, authBearer, authCookie } from '@pikku/core/middleware'
 export type { MiddlewarePriority }
 
 /**
