+++ packages/core/src/middleware/index.ts
@@ -3,7 +3,7 @@ export { authCookie } from './auth-cookie.js'
 export { authBearer } from './auth-bearer.js'
 export { pikkuRemoteAuthMiddleware } from './remote-auth.js'
 export { cors } from './cors.js'
-export { analyticsOrigin, isAllowedOrigin, toOrigin } from './analytics-origin.js'
+export { requireOrigin, isAllowedOrigin, toOrigin } from './require-origin.js'
 export { telemetryOuter, telemetryInner } from './telemetry.js'
 export {
   addTagMiddleware,
