+++ packages/core/src/public-surface.json
@@ -6,7 +6,6 @@
     "addGlobalPermission",
     "addMiddleware",
     "addTagMiddleware",
-    "analyticsOrigin",
     "authAPIKey",
     "authBearer",
     "authCookie",
@@ -18,6 +17,7 @@
     "pikkuMiddleware",
     "pikkuMiddlewareFactory",
     "pikkuRemoteAuthMiddleware",
+    "requireOrigin",
     "runMiddleware",
     "telemetryInner",
     "telemetryOuter",
